get proposal data by id
